Change to float

+++ bitget-python-sdk-api/bitget/bitget_api.py
@@ -7,8 +7,8 @@ class BitgetApi(Client):
     def __init__(self, api_key, api_secret_key, passphrase, use_server_time=False, first=False):
         Client.__init__(self, api_key, api_secret_key, passphrase, use_server_time, first)
 
-    def post(self, request_path, params, timeout: int | None = 30):
+    def post(self, request_path, params, timeout: float | None = 30):
         return self._request_with_params(POST, request_path, params, timeout)
 
-    def get(self, request_path, params, timeout: int | None = 30):
+    def get(self, request_path, params, timeout: float | None = 30):
         return self._request_with_params(GET, request_path, params, timeout)
